refactor(runtime): split addon runtime helpers
