+++ addons/realtime/realtime.go
@@ -5,7 +5,7 @@ package realtime
 import (
 	"github.com/cssbruno/gowdk"
 	"github.com/cssbruno/gowdk/runtime/contracts"
-	"github.com/cssbruno/gowdk/runtime/contracts/sse"
+	runtimerealtime "github.com/cssbruno/gowdk/runtime/realtime"
 )
 
 // ImportPath is the canonical Go import path for the realtime addon.
@@ -15,10 +15,10 @@ const ImportPath = "github.com/cssbruno/gowdk/addons/realtime"
 type PresentationFanout = contracts.PresentationFanout
 
 // SSEHub fans presentation events out to connected server-sent events clients.
-type SSEHub = sse.Hub
+type SSEHub = runtimerealtime.SSEHub
 
 // SSEOption configures a dependency-free SSE hub.
-type SSEOption = sse.Option
+type SSEOption = runtimerealtime.SSEOption
 
 // Addon enables realtime presentation-event fanout support.
 func Addon() gowdk.Addon {
@@ -27,10 +27,10 @@ func Addon() gowdk.Addon {
 
 // NewSSE creates a dependency-free server-sent events presentation fanout hub.
 func NewSSE(options ...SSEOption) *SSEHub {
-	return sse.New(options...)
+	return runtimerealtime.NewSSE(options...)
 }
 
 // WithSSEBufferSize sets each SSE client's queued message buffer.
 func WithSSEBufferSize(size int) SSEOption {
-	return sse.WithBufferSize(size)
+	return runtimerealtime.WithSSEBufferSize(size)
 }
