+++ runtime/partial/client_hooks.go
