+++ runtime/partial/swap.go
